# api.yaml
openapi: 3.0.0
info:
  version: 0.0.1
  title: My User Project CRUD
  description: My User Project Application API
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'http://localhost:3000/'
tags:
  - name: user
    description: Everything about users
paths:
  /users:
    get:
      tags:
        - user
      summary: Get All Users
      operationId: getAllUsers
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
        - user
      summary: Creates a user
      operationId: postUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - firstName
                - lastName
                - weight
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                weight:
                  type: integer
      responses:
        '201':
          description: User created
  /users/{id}:
    get:
      tags:
        - user
      summary: Get a user by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the user to get
      responses:
        '200':
          description: User
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags:
        - user
      summary: Update a user
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the user to get
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                weight:
                  type: integer
      responses:
        '200':
          description: User updated
    delete:
      tags:
        - user
      summary: Delete a user by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the user to get
      responses:
        '204':
          description: User deleted
    put:
      tags:
        - user
      summary: Update a user
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric ID of the user to get
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                weight:
                  type: integer
      responses:
        '200':
          description: User updated
components:
  schemas:
    User:
      type: object
      required:
        - firstName
        - lastName
        - weight
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        weight:
          type: integer
      example:
        id: 1
        firstName: Léo
        lastName: BONGIORNO
        weight: 62